+++ src/helpers/dashboards.ts
@@ -1,0 +1,36 @@
+import { Dashboard, DashboardConfig } from '../types';
+
+export const parseDashboards = (data) => {
+  const dashboards: Record<string, Dashboard> = {};
+  data.forEach((dashboard) => {
+    if (dashboard.mode == 'storage') {
+      dashboards[dashboard.id] = dashboard;
+    }
+  });
+  return dashboards;
+};
+
+export const parseDashboardGenerator = (dashboardId, dashboardUrl) => {
+  const func = async (dashboardConfig: DashboardConfig) => {
+    const response = {
+      templates: {},
+      dashboard: dashboardConfig,
+      views: {},
+      dashboardId,
+      dashboardUrl,
+    };
+    if (dashboardConfig.template) {
+      dashboardConfig.views.forEach((view) => {
+        if (view.cards?.length == 1) {
+          response.templates[`${view.path}`] = view.cards[0];
+        }
+      });
+    }
+    dashboardConfig.views.forEach((view) => {
+      response.views[`${dashboardId}${view.path ? `.${view.path}` : ''}`] = view;
+    });
+    dashboardConfig.views = Object.values(response.views);
+    return response;
+  };
+  return func;
+};
