+++ src/helpers/templates.ts
@@ -1,0 +1,21 @@
+import { DashboardCard, DashboardView } from '../types';
+
+export const getTemplatesUsedInCard = (card: DashboardCard): string[] => {
+  if (card.template) {
+    return [card.template];
+  }
+  if (card.cards) {
+    return card.cards.flatMap((c) => {
+      return getTemplatesUsedInCard(c);
+    });
+  }
+  return [];
+};
+
+export const getTemplatesUsedInView = (view: DashboardView): string[] => {
+  return (
+    view.cards?.flatMap((c) => {
+      return getTemplatesUsedInCard(c);
+    }) || []
+  );
+};
